+++ docs/pages/api-reference/_meta.tsx
@@ -1,3 +1,4 @@
 export default {
+  "index": "Overview",
   "data-providers": "Data Provider Interfaces"
 }